+++ pychron/paths.py
@@ -77,6 +77,7 @@ class Paths():
     plugins_dir = None
     hidden_dir = None
     preferences_dir = None
+    comment_templates_dir = None
     plotter_options_dir = None
     test_dir = None
     custom_queries_dir = None
@@ -231,6 +232,7 @@ class Paths():
         self.hidden_dir = join(root, '.hidden')
         self.preferences_dir = join(root, 'preferences')
         self.plotter_options_dir = join(self.hidden_dir, 'plotter_options')
+        self.comment_templates_dir = join(self.hidden_dir, 'comment_templates')
         # self.test_dir = join(root, 'testing')
         # self.custom_queries_dir = join(root, 'custom_queries')
         self.template_dir = join(root, 'templates')