+++ pychron/processing/plotters/options/plotter.py
@@ -28,7 +28,7 @@ from traitsui.api import View, Item, HGroup, VGroup, Group, \
 from traitsui.extras.checkbox_column import CheckboxColumn
 from traitsui.table_column import ObjectColumn
 from pychron.envisage.tasks.pane_helpers import icon_button_editor
-from pychron.processing.label_maker import TitleMaker
+from pychron.processing.label_maker import TitleTemplater
 from pychron.processing.plotters.options.base import FigurePlotterOptions
 from pychron.pychron_constants import ALPHAS
 
@@ -79,7 +79,7 @@ class PlotterOptions(FigurePlotterOptions):
     x_filter_str = Str
 
     def _edit_title_format_fired(self):
-        tm = TitleMaker(label=self.title,
+        tm = TitleTemplater(label=self.title,
                         delimiter=self.title_delimiter,
                         leading_text=self.title_leading_text,
                         trailing_text=self.title_trailing_text)